+++ frontend/src/features/informationTabs/questionTab/components/SubmitButton.module.css
@@ -1,0 +1,24 @@
+/* Base button styling */
+.baseButton {
+  padding: 10px 12px;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  font-weight: 500;
+  font-size: 0.95rem;
+  width: 100%;
+  transition: background-color 0.2s ease;
+}
+
+/* Submit Button */
+.submitButton {
+  margin-top: 1.5rem;
+  z-index: 9999;
+  background-color: #10b981;
+  color: white;
+}
+
+.submitButton:hover,
+.submitButton:active {
+  background-color: #059669;
+}
